client/requirements.txt
boto3==1.9.144
botocore==1.12.253
certifi==2020.6.20
chardet==3.0.4
docutils==0.15.2
idna==2.10
jmespath==0.10.0
python-dateutil==2.8.1
python-dotenv==0.10.1
requests==2.24.0
s3transfer==0.2.1
six==1.15.0
urllib3==1.25.10
